Implement fallback to IPv4/IPv6 vice-versa for sync
This commit changes the way the driver attempts connections to the
ip addresses that are associated with a host. Instead of only fetching
the first ip address and attempting a connection that single ip address,
the driver will now fetch all of the ip addresses and attempt to
connect to each one until a successful connection is made, or until
it is determined that none of the ip addresses can be connected to.
Note that this commit is only for sync. The implementation of this
for async will be handled in a separate commit. A couple of tests
were added for this change as well.

JAVA-2700